service id int -> int64

+++ internal/api/category/get.go
@@ -8,7 +8,7 @@ import (
 )
 
 func (categoryImpl *Implementation) Get(ctx context.Context, req *desc.GetRequest) (*desc.GetResponse, error) {
-	category, err := categoryImpl.categoryService.Get(ctx, int(req.Id))
+	category, err := categoryImpl.categoryService.Get(ctx, int64(req.Id))
 	if err != nil {
 		return nil, err
 	}
