+++ internal/repository/question_subcategory/pg/converter/question_subcategory.go
@@ -4,11 +4,11 @@ import (
 	modelRepo "github.com/Kosfedev/learn_go/internal/repository/question_subcategory/pg/model"
 )
 
-func QuestionSubcategoriesIDsFromPGSQL(questionSubcategoriesRepo []*modelRepo.QuestionSubcategory) []int {
-	questionSubcategories := make([]int, len(questionSubcategoriesRepo))
+func QuestionSubcategoriesIDsFromPGSQL(questionSubcategoriesRepo []*modelRepo.QuestionSubcategory) []int64 {
+	questionSubcategories := make([]int64, len(questionSubcategoriesRepo))
 
 	for i, questionSubcategoryRepo := range questionSubcategoriesRepo {
-		questionSubcategories[i] = int(questionSubcategoryRepo.SubcategoryID)
+		questionSubcategories[i] = int64(questionSubcategoryRepo.SubcategoryID)
 	}
 
 	return questionSubcategories
