+++ internal/repository/question_subcategory/pg/list_subcategories_by_question_id.go
@@ -8,7 +8,7 @@ import (
 	modelRepo "github.com/Kosfedev/learn_go/internal/repository/question_subcategory/pg/model"
 )
 
-func (r *repo) ListSubcategoriesByQuestionID(ctx context.Context, questionID int) ([]int, error) {
+func (r *repo) ListSubcategoriesByQuestionID(ctx context.Context, questionID int64) ([]int64, error) {
 	idRepo := int32(questionID)
 	query := db.Query{
 		Name:     "question_subcategory_repository.list_subcategories_by_question_id",
