+++ internal/repository/questionset/pg/converter/question_set.go
@@ -9,7 +9,7 @@ import (
 
 func QuestionSetFromPGSQL(questionSetRepo *modelRepo.QuestionSet) *model.QuestionSet {
 	questionSetServ := &model.QuestionSet{
-		ID:        int(questionSetRepo.ID),
+		ID:        int64(questionSetRepo.ID),
 		Name:      questionSetRepo.Name,
 		CreatedAt: questionSetRepo.CreatedAt,
 	}
