+++ internal/service/category/update.go
@@ -6,7 +6,7 @@ import (
 	"github.com/Kosfedev/learn_go/internal/model"
 )
 
-func (qcs *serv) Update(ctx context.Context, id int, updatedCategory *model.UpdatedCategory) error {
+func (qcs *serv) Update(ctx context.Context, id int64, updatedCategory *model.UpdatedCategory) error {
 	err := qcs.repo.Update(ctx, id, updatedCategory)
 	if err != nil {
 		return err
