+++ internal/service/domain/tests/create_test.go
@@ -15,7 +15,7 @@ import (
 func TestCreate(t *testing.T) {
 	t.Parallel()
 	ctx := context.Background()
-	id := int(gofakeit.Int64())
+	id := int64(gofakeit.Int64())
 	req := &model.NewDomain{
 		Name: gofakeit.Name(),
 	}
