+++ internal/service/question/add_subcategories.go
@@ -4,7 +4,7 @@ import (
 	"context"
 )
 
-func (qs *serv) AddSubcategories(ctx context.Context, questionID int, subcategoryIDs []int) error {
+func (qs *serv) AddSubcategories(ctx context.Context, questionID int64, subcategoryIDs []int64) error {
 	err := qs.questionSubcategoryRepo.AddSubcategoriesToQuestion(ctx, questionID, subcategoryIDs)
 	if err != nil {
 		return err
