+++ internal/service/question/create.go
@@ -7,7 +7,7 @@ import (
 	"github.com/Kosfedev/learn_go/internal/model"
 )
 
-func (qs *serv) Create(ctx context.Context, newQuestion *model.NewQuestion) (int, error) {
+func (qs *serv) Create(ctx context.Context, newQuestion *model.NewQuestion) (int64, error) {
 	if newQuestion == nil {
 		return 0, errors.New("newQuestion cannot be nil")
 	}
